% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGames.R
\name{runGames}
\alias{runGames}
\title{Team Win Data App}
\usage{
runGames()
}
\description{
This function allows the team win data shiny app to run. The app prompts the
user to type in a team name, showing suggestions if they don't input a valid
team name. Once the user inputs a valid team name, the app shows that team's
win and loss data as well as a summary chart displaying the same information.
}
