% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUU_performance.R
\name{SUU_performance}
\alias{SUU_performance}
\title{Create a quick scatter plot in ggplot.}
\usage{
SUU_performance()
}
\value{
This function returns a ggplot bar plot object.
}
\description{
This will graph SUU's performance based on the games2023 data set.
}
\examples{
## Create a barplot for SUU's wins and losses.
SUU_performance()

}
