% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_performance.R
\name{team_performance}
\alias{team_performance}
\title{Create a quick scatter plot in ggplot.}
\usage{
team_performance(team)
}
\value{
This function returns a ggplot bar plot object.
}
\description{
This will graph a given team's performance based on the games2023 data set.
}
\examples{
## Create a barplot for SUU's wins and losses.
team_performance("Southern Utah")

}
