% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_summary.R
\name{team_summary}
\alias{team_summary}
\title{Create a quick scatter plot in ggplot.}
\usage{
team_summary(df, team_name)
}
\arguments{
\item{df}{This is the data frame containing all information.}

\item{team_name}{This is the name of the team we want a summary of.}
}
\value{
This function returns a single tibble row for the team with
name, number of wins, losses, and win percentage for columns.
}
\description{
This will summarize a given team's performance from a data set containing game data.
}
\examples{
## Summary data for SUU.
data <- basketPlots::games2023
team_summary(data, "Southern Utah")

}
