% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_loss_scatter.R
\name{win_loss_scatter}
\alias{win_loss_scatter}
\title{Create a scatter plot of the teams wins vs losses.}
\usage{
win_loss_scatter()
}
\value{
This function returns an interactive ggplot scatter plot object.
}
\description{
This will graph the performance of all teams by creating an interactive ggplot style
scatter plot of the teams wins vs losses through plotly. The data comes from the winData data set.
}
\examples{
## Create a scatter plot of the teams wins vs losses.
win_loss_scatter()

}
